+++ src/api/handcash_connect_api_error.ts
@@ -1,4 +1,6 @@
 export default class HandCashConnectApiError extends Error {
+	message: string;
+
 	path: string;
 
 	method: string;